requirements.txt
absl-py==1.4.0
accelerate==0.21.0
ai4bharat-transliteration==1.1.3
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
altair==5.0.1
antlr4-python3-runtime==4.8
anyascii==0.3.2
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
backports.zoneinfo==0.2.1
bangla==0.0.2
bitarray==2.8.0
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
boltons==23.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.6
clip-anytorch==2.5.2
cmake==3.27.0
colorama==0.4.6
contourpy==1.1.0
coqpit==0.0.17
cycler==0.11.0
Cython==0.29.28
dateparser==1.1.8
decorator==5.1.1
dill==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.18.1
einops==0.6.1
fairseq==0.12.2
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
g2pkk==0.1.2
gast==0.4.0
gevent==23.7.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h5py==3.9.0
huggingface-hub==0.16.4
hydra-core==1.0.7
idna==3.4
imageio==2.31.1
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
indic-nlp-library==0.92
inflect==5.6.0
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.1
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
k-diffusion==0.0.15
keras==2.13.1
kiwisolver==1.4.4
kornia==0.6.12
lazy-loader==0.3
libclang==16.0.6
librosa==0.10.0.post2
lit==16.0.6
llvmlite==0.38.1
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mecab-python3==1.0.5
mock==5.1.0
Morfessor==2.0.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.55.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.0.6
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
pooch==1.6.0
portalocker==2.7.0
progressbar2==4.2.0
promise==2.3
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.0
pydload==1.0.9
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pynndescent==0.5.10
pyparsing==3.0.9
pypinyin==0.49.0
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-utils==3.7.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rich==13.5.0
rpds-py==0.9.2
rsa==4.9
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soundfile==0.12.1
soxr==0.3.5
sphinx==6.2.1
sphinx-argparse==0.4.0
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
streamlit==1.25.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
tensorflow==2.13.0
tensorflow-addons==0.21.0
tensorflow-datasets==3.2.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
termcolor==2.3.0
tf2crf==0.1.33
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchsde==0.2.5
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
trainer==0.0.20
trampoline==0.1.2
transformers==4.31.0
triton==2.0.0
TTS==0.14.3
typeguard==2.13.3
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
umap-learn==0.5.1
unidic-lite==1.0.8
urduhack==1.1.1
urllib3==2.0.4
validators==0.20.0
wandb==0.15.7
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
